+++ src/heretic/p_floor.cpp
@@ -203,7 +203,7 @@ void T_MoveFloor(floormove_t * floor)
             switch (floor->type)
             {
                 case donutRaise:
-                    floor->sector->special = floor->newspecial;
+                    floor->sector->special = static_cast<short>(floor->newspecial);
                     floor->sector->floorpic = floor->texture;
                 default:
                     break;
@@ -212,7 +212,7 @@ void T_MoveFloor(floormove_t * floor)
             switch (floor->type)
             {
                 case lowerAndChange:
-                    floor->sector->special = floor->newspecial;
+                    floor->sector->special = static_cast<short>(floor->newspecial);
                     floor->sector->floorpic = floor->texture;
                 default:
                     break;
@@ -429,11 +429,11 @@ int EV_BuildStairs(line_t * line, fixed_t stepDelta)
                     continue;
 
                 tsec = (sec->lines[i])->frontsector;
-                newsecnum = tsec - sectors;
+                newsecnum = static_cast<int>(tsec - sectors);
                 if (secnum != newsecnum)
                     continue;
                 tsec = (sec->lines[i])->backsector;
-                newsecnum = tsec - sectors;
+                newsecnum = static_cast<int>(tsec - sectors);
                 if (tsec->floorpic != texture)
                     continue;
