+++ src/heretic/p_map.cpp
@@ -882,7 +882,7 @@ bool P_TryMove(mobj_t * thing, fixed_t x, fixed_t y)
             if (side != oldside)
             {
                 if (ld->special)
-                    P_CrossSpecialLine(ld - lines, oldside, thing);
+                    P_CrossSpecialLine(static_cast<int>(ld - lines), oldside, thing);
             }
         }
