+++ src/hexen/r_main.cpp
@@ -606,7 +606,7 @@ void R_ExecuteSetViewSize()
 // thing clipping
 //
     for (i = 0; i < viewwidth; i++)
-        screenheightarray[i] = viewheight;
+        screenheightarray[i] = static_cast<short>(viewheight);
 
 //
 // planes
